Refactor deep copy implementation and introduce CopyTo methods for Person and Employee classes
