+++ DesignPatterns.Prototype/Inheritance/Address.cs
@@ -1,0 +1,25 @@
+namespace DesignPatterns.Prototype.Inheritance;
+
+public class Address : IDeepCopyable<Address>
+{
+    public string StreetName;
+    public int HouseNumber;
+
+    public Address() { }
+    public Address(string streetName, int houseNumber)
+    {
+        StreetName = streetName;
+        HouseNumber = houseNumber;
+    }
+
+    public void CopyTo(Address target)
+    {
+        target.StreetName = StreetName;
+        target.HouseNumber = HouseNumber;
+    }
+
+    public override string ToString()
+    {
+        return $"{nameof(StreetName)}: {StreetName}, {nameof(HouseNumber)}: {HouseNumber}";
+    }
+}
